Attempt to bring some sanity to process & thread models.
ProcessModel now contains one row per pid+thread, and is the resource for
accessing information about both the threads and the process they belong to.

The expectation is that section headers can be used to present this information
in a useful way.

ThreadModel will contain slices (as in actual data) about activity on the
threads.